+++ argilla-server/CHANGELOG.md
@@ -19,7 +19,7 @@ These are the section headers that we use:
 
 ### "Added"
 - Pytest and python env setup in GH Codespaces devcontainer
-
+- Added Response update duration tests
 
 ### "Changed"
 - Updated elasticsearch to 8.15.0